// program/clients/typescript/src/generated/pdas/merchantOperatorConfig.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  getAddressEncoder,
  getProgramDerivedAddress,
  getU32Encoder,
  getUtf8Encoder,
  type Address,
  type ProgramDerivedAddress,
} from '@solana/kit';

export type MerchantOperatorConfigSeeds = {
  merchant: Address;
  operator: Address;
  version: number;
};

export async function findMerchantOperatorConfigPda(
  seeds: MerchantOperatorConfigSeeds,
  config: { programAddress?: Address | undefined } = {}
): Promise<ProgramDerivedAddress> {
  const {
    programAddress = 'commkU28d52cwo2Ma3Marxz4Qr9REtfJtuUfqnDnbhT' as Address<'commkU28d52cwo2Ma3Marxz4Qr9REtfJtuUfqnDnbhT'>,
  } = config;
  return await getProgramDerivedAddress({
    programAddress,
    seeds: [
      getUtf8Encoder().encode('merchant_operator_config'),
      getAddressEncoder().encode(seeds.merchant),
      getAddressEncoder().encode(seeds.operator),
      getU32Encoder().encode(seeds.version),
    ],
  });
}